refactor: node optimization
